requirements.txt
pysnmp==4.4.12
PyYAML==5.2